Deploy 201

+++ Backend_UMR_Work_Program/DataModels/WP_RESERVES_REPLACEMENT_RATIO_VALUE_TO_BE_PIVOTTED.cs
@@ -6,7 +6,7 @@ namespace Backend_UMR_Work_Program.Models
     public partial class WP_RESERVES_REPLACEMENT_RATIO_VALUE_TO_BE_PIVOTTED
     {
         public string? CompanyName { get; set; }
-        public int? RESERVES_REPLACEMENT_RATIO_VALUE { get; set; }
+        public decimal? RESERVES_REPLACEMENT_RATIO_VALUE { get; set; }
         public string? Year_of_WP { get; set; }
     }
 }
